feat(suite): move Address component to suite package
