+++ packages/components/src/index.ts
@@ -5,7 +5,6 @@ export { motionAnimation, motionEasing } from './config/motion';
 export { Checkbox, type CheckboxProps } from './components/form/Checkbox/Checkbox';
 export * from './components/animations/DeviceAnimation';
 export * from './components/animations/LottieAnimation';
-export { Address, type AddressProps } from './components/Address/Address';
 export { AssetLogo, type AssetLogoProps } from './components/AssetLogo/AssetLogo';
 export * from './components/Flag/Flag';
 export * from './components/AutoScalingInput/AutoScalingInput';
